Basic import of JSON bookmarks
This handles a very basic import of JSON formatted bookmarks.
The UI allows you to navigate to a folder in your bookmarks
and then enter JSON (paste is easiest) into a text area and
import it at that level.

This is functional for an MVP release but I would like to try
and make the UI use its own tab in the future. I think that would
allow me to use a file input to select a JSON file.