Use Elm for main UI
This switches over to using Elm for the main UI of the app.
The background script is still done in JS since it relies
heavily on Chrome APIs. The main interaction of traversing
and creating bookmarks is now done with Elm though. This uses
ports to access Chrome APIs for actual bookmark querying and
creation.

It ends up with a pretty cool separation. A JS layer/API to
interact with bookmark data and then pure data inside of Elm
to actually interact with the data.

This also brings in a much more formal build that includes js,
elm, html, and css. Also includes a watch script for development.