Rebuilds on top of TS and React
This is a rebuild of the elm app on top of TS and React.
I kept the UX more or less the same for this iteration.
The main goal was to use my current starter with TS and React to
recreate the same functionality.

I have some ideas for how the UX could change in the future, but wanted
to get this built first.